feat: создает модульную структуру
